chatbot/requirements.txt
python-dotenv==1.1.1
torch==2.6.0
torch==2.7.1
torch_xla==2.7.0
tqdm==4.65.0
tqdm==4.67.1
transformers==4.54.1
sentence_transformers==5.0.0
langchain==0.3.27
langchain_google_genai==2.1.9
langchain_mongodb==0.6.2
pymongo==4.13.2
Django==5.2.4